made route for getting container stream by id
