+++ Client/Containers/Navbar.tsx
@@ -1,16 +1,19 @@
 import * as React from 'react';
 import { Link, Route, Routes } from 'react-router-dom';
+import axios from 'axios';
 import Homepage from '../Components/Homepage';
 import DataOverview from './DataOverview';
 import YourContainers from './YourContainers';
 import Team from '../Components/Team';
 
 export default function Navbar() {
-  const sse = new EventSource('http://localhost:3000/cont/constream');
-  sse.onmessage = (event) => {
-    console.log(JSON.parse(event.data))
-  }
+  axios('cont/list')
+    .then(res => console.log(res.data))
 
+    const sse = new EventSource('http://localhost:3000/cont/constream/?id=9aac17eb0059');
+    sse.onmessage = (event) => {
+      console.log(JSON.parse(event.data))
+    }
 
   return (
     <div className="flex flex-col items-center text-3xl font-bold underlined bg-blue-800 w-1/5 max-w-xs">
