+++ server/routes/ContainerRoute.js
@@ -1,74 +1,18 @@
 const express = require('express');
+const { RouterProvider } = require('react-router-dom');
 const router = express.Router();
-const util = require('util');
-const { spawn }  = require('child_process');
-const exec = util.promisify(require('child_process').exec);
-
-//const dockerStat = spawn('docker', ['stats', '--format="{{json .}}"']);
+const containerController = require('../controllers/ContainerController');
 
 router.get('/', (req, res) => {
   res.send('hello from container route');
 });
 
+router.get('/constream/', containerController.dockerStatRequest);
 
+//router.get('/constream/', containerController.dockerStatRequestById);
 
-const dockerStatRequest = async (req, res, next) => {
-  const writeStats = (data) => {
-    data = JSON.stringify(data);
-    res.write('data: ' + data + '\n\n');
-  };
-  const parseData = (stdout) => {
-    const containers = [];
-    const dockerStats = stdout.trim();
-  
-    const conts = dockerStats.split('\n');
-  
-    for (let i = 0; i < conts.length; i++) {
-      containers.push(JSON.parse(conts[i]));
-    }
-    return containers;
-  }
-  res.writeHead(200, {
-    'Content-Type': 'text/event-stream',
-    'Cache-Control': 'no-cache',
-    'Connection': 'keep-alive',
-  });
-  try {
-    const { stdout } = await exec('docker stats --no-stream --format "{{json .}}"');
-    console.log(stdout);
-    const newData = parseData(stdout);
-    setInterval(() => writeStats(newData), 1000);
-  }
-  catch(err) {
-    next(err);
-  }
-      
-};
-
-router.get('/constream', dockerStatRequest);
-
-
-// router.get('/constream', async (req, res, next) => {
-//     //code that will grab the info from the docker cli
-//     try {
-//         dockerStat.stdout.on('data', data => {
-//             console.log(`${data}`);
-//             res.write(`data: ${data} \n\n`);
-//         })
-
-//         dockerStat.stderr.on('data', data => {
-//             console.log(`stderr: ${data}`);
-
-
-//     }
-//     catch(err) {
-//         next(err);
-//     }
-//     //await exec function reading the CLI following the input of `docker stats`
-//     //store this in a variable ?
-//         //send variable with data to front end
-//         //this function needs to call itself since it is part of the stream event 
-    
-// });
+router.get('/list', containerController.dockerContainers, (req, res) => {
+  res.status(200).json(res.locals.containers);
+});
 
 module.exports = router;
